detr3d ttcnn functional
