+++ models/experimental/detr3d/README.md
@@ -20,7 +20,7 @@ DETR3D is a transformer-based approach for 3D object detection that extends the 
 ```
 models/experimental/detr3d/
 ├── demo/                          # Demo and inference scripts
-│   └── detr3d_demo.py            # Main demo script with AP calculation
+│   └── demo.py            # Main demo script with AP calculation
 ├── reference/                     # PyTorch reference implementation
 │   ├── model_3detr.py            # Main DETR3D model implementation
 │   ├── model_config.py            # Model configuration classes
@@ -79,7 +79,7 @@ test_dataset = datasets["test"]
 Run the demo script for inference with AP calculation:
 
 ```bash
-python models/experimental/detr3d/demo/detr3d_demo.py \
+python models/experimental/detr3d/demo/demo.py \
     --dataset-root-dir /path/to/sunrgbd/dataset \
     --test-ckpt /path/to/checkpoint.pth \
     --seed 0
@@ -93,7 +93,7 @@ models/experimental/detr3d/resources/detr3d_weights_download.sh
 ```
 
 ```bash
-python models/experimental/detr3d/demo/detr3d_demo.py \
+python models/experimental/detr3d/demo/demo.py \
     --dataset-root-dir  models/experimental/detr3d/resources/sunrgbd/
     --test-ckpt models/experimental/detr3d/resources/sunrgbd_masked_ep720.pth
 ```
