+++ models/experimental/detr3d/common.py
@@ -30,7 +30,7 @@ def load_partial_state(torch_model: torch.nn.Module, state_dict, layer_name: str
 
 
 def load_torch_model_state(torch_model: torch.nn.Module = None, layer_name: str = "", model_location_generator=None):
-    if model_location_generator is None or "TT_GH_CI_INFRA" not in os.environ:
+    if model_location_generator == None or "TT_GH_CI_INFRA" not in os.environ:
         model_path = "models"
     else:
         model_path = model_location_generator("vision-models/detr3d", model_subdir="", download_if_ci_v2=True)
