+++ models/experimental/detr3d/reference/pointnet2_modules.py
@@ -1,5 +1,7 @@
 # SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
 # SPDX-License-Identifier: Apache-2.0
+
 # Copyright (c) Facebook, Inc. and its affiliates.
 """
 Modified from https://github.com/facebookresearch/3detr
