+++ models/experimental/detr3d/reference/utils/dataset/sunrgbd.py
@@ -1,5 +1,7 @@
 # SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
 # SPDX-License-Identifier: Apache-2.0
+
 # Modified from the 3Detr implementation in facebookresearch
 # https://github.com/facebookresearch/3detr
 # Copyright (c) Facebook, Inc. and its affiliates.
