+++ models/experimental/detr3d/resources/detr3d_weights_download.sh
