+++ models/experimental/detr3d/ttnn/pointnet_samodule_votes.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
 # SPDX-License-Identifier: Apache-2.0
 
@@ -170,6 +170,10 @@ class TtnnFurthestPointSampling(LightweightModule):
 
         # Initialize farthest indices
         farthest = ttnn.zeros((B,), dtype=ttnn.int32, layout=ttnn.TILE_LAYOUT, device=device)
+
+        # Create batch indices
+        batch_indices = ttnn.arange(0, B, dtype=ttnn.int32, device=device)
+        batch_indices = ttnn.reshape(batch_indices, (B, 1))
 
         centroid_list = []
         for i in range(n_samples):
